Create Clone-Man
Create a new Enemy called "Clone Man".
He should be having the following states:
Idle
Patrolling
Chasing
Shooting
Teleporting
Cloning
When this enemy is killed by the player, it creates 2 clones of itself with full health.
the 2 new cloned enemies will be teleported to some random place upon creation and enter the chasing state.
The cloned enemies should be visibly different in the game, highlight them with a different color.
Each original Clone Man will be able to clone twice only. The code should be flexible enough to increase or decrease this number as per requirements.
Create a CloneManStateMachine. Use the generic state machine for handling state management for Clone Man.
Create a CloneManController to represent your new enemy.
Create a new Level and add multiple Clone Man in this Level.
Share the link of the correct branch as submission and a gameplay video of your game as well