Create Coins and Scoring System
Create currency in your game in form of Coins
Whenever an enemy dies, some coins will be spilled on that location like below:
Player will be able to collect these coins when standing close enough to these coins.
The Coins you collect will contribute to our overall score.
The coins should have a pull animation -> When the player is in close proximity of a coin scattered around, that coin will be pulled towards the player position and player score will be increased.